refactor generator commands
