+++ src/Commands/GenerateControllerCommand.php
@@ -4,10 +4,8 @@
 namespace Experus\Sockets\Commands;
 
 use Experus\Sockets\Contracts\Kernel;
-use Illuminate\Console\Command;
-use Illuminate\Contracts\Foundation\Application;
 
-class GenerateControllerCommand extends Command
+class GenerateControllerCommand extends AbstractGeneratorCommand
 {
     /**
      * The name and signature of the console command.
@@ -22,70 +20,18 @@ class GenerateControllerCommand extends Command
      * @var string
      */
     protected $description = 'Generate a sockets controller class.';
-    /**
-     * @var Application
-     */
-    private $app;
 
     /**
-     * Create a new command instance.
+     * The default namespace to generate into.
      *
-     * @param Application $app
+     * @var string
      */
-    public function __construct(Application $app)
-    {
-        parent::__construct();
-        $this->app = $app;
-    }
+    protected $namespace = Kernel::CONTROLLER_NAMESPACE;
 
     /**
-     * Execute the console command.
-     */
-    public function handle()
-    {
-        $name = $this->argument('name');
-        $namespace = (is_null($this->option('namespace')) ? Kernel::CONTROLLER_NAMESPACE : $this->option('namespace'));
-        $directory = $this->buildDirectory($namespace);
-        $output = $directory . '/' . $name . '.php';
-
-        if (!is_dir($directory)) {
-            mkdir($directory, 0755, true);
-        }
-
-        $template = file_get_contents(__DIR__ . '/../../files/templates/Controller');
-        $template = $this->compile($template, compact('name', 'namespace'));
-
-        file_put_contents($output, $this->compile($template));
-    }
-
-    /**
-     * Get the directory to generate the controller in.
+     * The template file to generate from.
      *
-     * @param $namespace
-     * @return string
+     * @var string
      */
-    private function buildDirectory($namespace)
-    {
-        $base = is_null($this->option('dir')) ? 'app' : $this->option('dir');
-
-        $namespace = strstr($namespace, '\\');
-
-        return $this->app->basePath() . '/' . $base . str_replace('\\', '/', $namespace);
-    }
-
-    /**
-     * Compile a template file using the given environment variables.
-     *
-     * @param string $template
-     * @param array $env
-     * @return string
-     */
-    private function compile($template, array $env = [])
-    {
-        foreach ($env as $name => $value) {
-            $template = str_replace('{{' . $name . '}}', $value, $template);
-        }
-
-        return preg_replace('/{{[a-zA-Z]+}}/', '', $template);
-    }
+    protected $template = 'Controller';
 }
